Python_Data_Analysis/연습용디렉토리/requirements.txt
certifi==2019.9.11
chardet==3.0.4
collections-extended==1.0.3
Django==2.2.7
django-appconf==1.0.3
django-extensions==2.2.5
django-imagekit==4.0.2
idna==2.8
JPype1==0.7.0
konlpy==0.5.1
nltk==3.4.5
pilkit==2.0
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.7
Wikipedia-API==0.5.3
